omx(team): checkpoint worker-1 shutdown changes
